Add draggable element to fired drag events.
It is not always reliable to use the event itself to find the draggable base.
In testing in Chrome, event.currentTarget was always window.
event.target was always the clicked element, which may not always be
the draggable instance if it has children.

The element passed to the event ensures that you always have access to
the base element if you need it.