Fix for 'cannot access clientX on undefined' exception.
To reproduce, access a DraggableCore element using touch emulation in Chrome.   The exception will be thrown on drag stop.

This will leave the element active, so if you drag another element, both elements end up being dragged together.

I'm not sure if this is the correct fix, I may be treating symptoms rather than root cause, but it appears to work.