feat(React): add `nodeRef` prop
If running in React Strict mode, ReactDOM.findDOMNode() is deprecated.
Unfortunately, in order for <Draggable> to work properly, we need raw access
to the underlying DOM node. If you want to avoid the warning, pass a `nodeRef`
as in this example:

function MyComponent() {
  const nodeRef = React.useRef(null);
  return (
    <Draggable nodeRef={nodeRef}>
      <div ref={nodeRef}>Example Target</div>
    </Draggable>
  );
}

This can be used for arbitrarily nested components, so long as the ref ends up
pointing to the actual child DOM node and not a custom component.

Thanks to react-transition-group for the inspiration.
`nodeRef` is also available on <DraggableCore>.